apps/graph/requirements.txt
fastapi[all]==0.105.0
uvicorn==0.20.0
python-dotenv==0.21.1
